🐛 fix(components): complete the interrupted tooltip refactor
The previous commit was split by lint-staged partial staging and
landed without the follow-up pieces, leaving stale import paths and
a missing Tooltip module.

- fix Volume import paths after the directory move
- add the generic Tooltip primitive (components/Tooltip)
- consume it from ProgressTooltip instead of the deleted local CSS
