+++ package/src/components/Tooltip/Tooltip.css
@@ -1,0 +1,16 @@
+.rmap-tooltip {
+  position: absolute;
+  bottom: calc(100% + 6px);
+  transform: translateX(-50%);
+  max-width: none;
+  padding: 2px 6px;
+  border-radius: 4px;
+  background-color: var(--rm-audio-player-text-color);
+  color: var(--rm-audio-player-interface-container);
+  font-size: 12px;
+  line-height: 1.4;
+  white-space: nowrap;
+  box-shadow: 0 2px 4px rgb(var(--rm-audio-player-shadow, 0 0 0) / 10%);
+  pointer-events: none;
+  z-index: 6;
+}
